WIP ctree,ioctl,utils: Add Block Group Tree
When the block_group_tree feature is enabled on a filesystem, we have to
look up Block Group metadata items in the new Block Group Tree, instead
of the Extent Tree, where they always were located before.

WIP: We don't want to call the get_features ioctl *every* time we look
up 1 block group object, but we also cannot just cache the result of it,
since some feature flags can change while the fs is mounted.

So, this needs some extra work in the features department first to make
this more nice.